moved rest of the colors and refactored theme file
