+++ packages/components/themes/fonts.tsx
@@ -6,3 +6,11 @@ import '@fontsource/montserrat/600.css'
 import '@fontsource/montserrat/700.css'
 import '@fontsource/montserrat/800.css'
 import '@fontsource/montserrat/900.css'
+
+export const typography = {
+  fontFamily: 'Montserrat',
+  fontWeightLight: 400,
+  fontWeightRegular: 500,
+  fontWeightMedium: 600,
+  fontWeightBold: 700,
+}
